refactor(raijin): separate project interfaces
