+++ yarn/raijin/src/project/index.ts
@@ -1,3 +1,4 @@
-export * from './manifest.js'
-export * from './type.js'
-export * from './workspaces.js'
+export *      from './manifest.js'
+export type * from './project.interfaces.js'
+export *      from './type.js'
+export *      from './workspaces.js'
