+++ yarn/raijin/src/project/manifest.ts
@@ -1,8 +1,4 @@
-export interface ProjectManifestSource {
-  workspaceDefinitions: ReadonlyArray<{
-    pattern: string
-  }>
-}
+import type { ProjectManifestSource } from './project.interfaces.js'
 
 export const getManifestWorkspacePatterns = (manifest: ProjectManifestSource): Array<string> =>
   manifest.workspaceDefinitions.map(({ pattern }) => pattern)
