+++ yarn/raijin/src/project/workspaces.ts
@@ -1,28 +1,9 @@
-import type { ProjectManifestSource }   from './manifest.js'
-import type { ProjectType }             from './type.js'
+import type { ProjectSource }           from './project.interfaces.js'
+import type { ProjectWorkspaceSource }  from './project.interfaces.js'
+import type { RaijinProjectModel }      from './project.interfaces.js'
 
 import { getManifestWorkspacePatterns } from './manifest.js'
 import { resolveProjectType }           from './type.js'
-
-export interface ProjectWorkspaceSource<TCwd extends string = string> {
-  cwd: TCwd
-  manifest: ProjectManifestSource
-}
-
-export interface ProjectSource<TWorkspace extends ProjectWorkspaceSource = ProjectWorkspaceSource> {
-  topLevelWorkspace: TWorkspace
-  workspaces: Iterable<TWorkspace>
-}
-
-export interface RaijinProjectModel<
-  TWorkspace extends ProjectWorkspaceSource = ProjectWorkspaceSource,
-> {
-  cwd: TWorkspace['cwd']
-  topLevelWorkspace: TWorkspace
-  type: ProjectType
-  workspacePatterns: Array<string>
-  workspaces: Array<TWorkspace>
-}
 
 export const createProjectModel = <TWorkspace extends ProjectWorkspaceSource>(
   project: ProjectSource<TWorkspace>
